Renaming build_server script, so certain functions can be imported elsewhere (coming soon)
